Create separate files for every tag
These are to be populated with any post which has a specific tag.

Given posts/gamedev.html, any post which includes the tag 'gamedev' must
be listed on this page in reverse chronological order.

If a post has tags 'a' and 'b' it must appear on both posts/a.html and
posts/b.html.